Kalista: Disable EFS
Only input from the BJ port is wired to VSYS on Kalista. VBUS from
USB-C is for output only. In other words, Kalista is a source only
device from a USB/PD perspective.

This patch disables EC-EFS, which would be needed for the EC to jump
to RW to get PD power before the AP boots. Kalist will be always
supplied enough power to boot the AP through the BJ port.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

CQ-DEPEND=CL:1341408
BUG=b:118386511
BRANCH=none
TEST=Boot Fizz. Verify normal boot, soft sync, recovery mode work.

Change-Id: I9efa9687868ab4f1e138237a10c244ce14955f13