gsctool: explicitly set buffering type to line buffered
Program stdout buffering type by default changes from line buffered to
block buffered if the output is redirected to a file or pipe. Since we
are going to call gsctool from inside debugd (CL:1337190) and want to
pipe the output instantly to the dbus request sender, the buffering type
of gsctool needs to be explicitly set.

BRANCH=none
BUG=b:113893821
TEST=run gsctool -O verify_ro_ZZAF.db from debugd (see CL:1337190)

Signed-off-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Change-Id: I515854a29e5e2ede0acc8c2d9e2c4df367a5062e