Update patch set 1
Patch Set 1:

That is, it must follow the rules for inclusion of third party code including a LICENCE file in the third_party directory.

Does this need to be forked from Linux, or can we just have it installed somewhere else in the chroot and call it if present?

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>