touchpad_st: do not generate 0 finger hid event
When we receive a beacon, we will also try to collect finger events.
In this case, there might be the case that there is not finger but dome
switch is on.  In this case, we should not report the click.

BRANCH=nocturne
BUG=b:119597909
TEST=manual on whiskers
Signed-off-by: Wei-Han Chen <stimim@chromium.org>

Change-Id: I03fe4481d17f6e919ab9501b2a93fa19635e381f
Reviewed-on: https://chromium-review.googlesource.com/1337253
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
(cherry picked from commit 65b8dcf6da612c774d372c3c29e16923d7f6f379)