pd_protocol: Don't enter LPM if port partner is connected.
The PD state machine can now put the TCPCs into low power mode in the
SNK/SRC_DISCONNECTED states in addition to the DRP_AUTO_TOGGLE state.
In order to prevent waking up the TCPCs, only CC status changes wake the
tasks when in this state.

This commit fixes a bug where the PD task would be suspended in the
SNK_DISCONNECTED state even though there was a charger plugged in.  If a
port partner is connected, the port should not enter low power mode.

BUG=b:119620277
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne, boot to S0, plug in zinger, blackcat, and guppy,
and verify that they all charge.
TEST=Verify that port goes into low power mode when the charger is
removed.
TEST=Verify that those chargers charge the system in all power states.

Change-Id: I54b613976d3709961f72c80da3acdd773cfe901e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>