mkbp: Enable the EC to report whether it has more events on mkbp_get_next_event
On certain platforms (e.g. nocturne), the EC --> AP MKBP synchronization
is handled by means of a level IRQ. This means that the EC will keep the
interrupt pin set as long as there are events to be served, but the AP
will need to re-enter its IRQ handler once per event in order to serve
all the events in the FIFO.

This commit adds a version 2 of EC_CMD_GET_NEXT_EVENT, such that the EC
will use the most-significant bit of the event type to record the fact
that the EC has more events available. This, in turn, enables the AP to
keep its interrupt handler thread awake and loop until all events
are served.

Since it uses a new command version, this change is forward and backward
compatible:
 - new EC, old kernel: the old kernel will use the V1 command
   and never see the flag
 - new kernel, old EC: the old EC will not accept the V2 command
   and never send the flag

BUG=b:119570064, b:112102925
TEST=patched Linux kernel can see and use the flag on nocturne
BRANCH=nocturne

Change-Id: I5bae7fdc85efcd26f7bdebcd31a7f27ecf570d88
Signed-off-by: Enrico Granata <egranata@chromium.org>