mkbp: Add a "more events" flag to the host event format
On certain platforms (e.g. nocturne), the EC --> AP MKBP synchronization
is handled by means of a level IRQ. This means that the EC will keep the
interrupt pin set as long as there are events to be served, but the AP
will need to re-enter its IRQ handler once per event in order to serve
all the events in the FIFO.

This commit uses the most-significant bit of the event type to record
the fact that the EC has more events available. This, in turn, enables
the AP to keep its interrupt handler thread awake and loop until all
the events are served.

Since this is a breaking change to the MKBP format, it is meant to only
initially be merged to the nocturne branch, and requires a corresponding
kernel change.

BUG=b:119570064
TEST=Linux kernel can see and use the flag on nocturne
BRANCH=nocturne

Change-Id: I5bae7fdc85efcd26f7bdebcd31a7f27ecf570d88
Signed-off-by: Enrico Granata <egranata@chromium.org>