Update patch set 1
Patch Set 1:

Is it necessary for this to be an entirely new file, versus changing the config on the baseboard file?  Also, does there need to be consideration for the board version here since protos do not have this chip?

Patch-set: 1
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>