lsm6dsm: Increase EC max sensor frequency
For most boards, the EC max sensor frequency is set to 200 Hz. When that
frequency is requested for boards using the lsm6dsm driver, the
ODR is rounded up to 208 Hz, which is then considered to be an invalid
parameter. The EC max sensor frequency has been increased to 208 Hz for
boards using the lsm6dsm driver.

BUG=b:118205424
BUG=b:118851581
TEST=Use ectool to set ODR to 208000 mHz for base accelerometer
TEST=Use ectool to set ODR to 208000 mHz for gyroscope
BRANCH=none

Change-Id: I97bcd83b6acfd3e17b2adb557a3216da3b0c5b98
Signed-off-by: Jesse Schettler <jschettler@chromium.org>