Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > I just added a comment to CL:1347015:
> > 
> > 
> > What if we added a condition to [0] that checked if odr == EX_MAX, then set roundup to false? That would cover add the scenarios that I can think of
> > 
> > [0] https://chromium.git.corp.google.com/chromiumos/platform/ec/+/9ba8fc3405d362ef503311560c21dcfcdfe67030/common/motion_sense.c#269
> 
> Assuming max_frequency is correct, when odr==max_frequency, round up is no op. 
> CTS tests expect frequencies to be rounded up, it can handle samples at 1.8x the asked frequency.

for the check I meant that the new odr == max_ec_rate (not sensor_max_rate). If we see that condition, then we know there was potential clipping at [0]. If we clipped, then we should round down instead of up. If there was not clipping, then rounding won't matter.

[0] https://chromium.googlesource.com/chromiumos/platform/ec/+/aa4b25ae7872e1a210060848cf1dbc773146e9ac/common/motion_sense.c#1161

Patch-set: 2