Update patch set 1
Patch Set 1:

I have this change in the kernel if you want it for testing:
https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1330058

Stephen thinks we should do this a different way, but we could also put that change in for now and then redo it if this is blocking you.

I guess you still need firmware support to keep it high before Linux comes up.

Patch-set: 1
CC: Gerrit User 1260470 <1260470@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>