cheza: Only one port can mux its DisplayPort lines
This is the limitation of the SoC, only one port can be its DisplayPort
line at a time. When a HPD happens, it checks if other port is already
muxed the DisplayPort lines.

It also configures the GPIO for enabling the port 1 redriver.

BRANCH=none
BUG=b:120142369
TEST=Plugged HDMI cable to a dongle to port 0, checked muxed to DP and
correct GPIO setting; then plugged HDMI cable to a dongle to port 1,
checked the port 1 not muxed to DP. Swapped the ports and did the same
check.

Change-Id: I06271b293f83250fb36b6958e09c8fb81e7122bc
Signed-off-by: Wai-Hong Tam <waihong@google.com>