PD: Ensure SVID parsing doesn't exceed packet boundaries
When responding to DiscSVID requests, most port partners will include an
SVID value of 0 to indicate the end of the SVIDs.  However, not all
dongles follow this rule and their packet may end with a VDO containing
two non-zero SVIDs.  In this scenario, we should stop parsing when we
reach the end of the valid packet region to avoid finding junk SVIDs.

BRANCH=octopus
BUG=b:116764439
TEST=plugged in several dongles with 1 or 2 VDOs each and verified we
parsed their SVIDs correctly

Change-Id: Id2b40ba918439aeb214efb43837dd995ff0b9d86
Signed-off-by: Diana Z <dzigterman@chromium.org>