ec3po: use a copy of the logger object
Use a copy of the console and interpreter logger objects, so modifying
the log level for one console does not affect the others.

With this change you can enable the debug log level on the cr50 console
without enabling debug for the ec console. The ec console prints a lot
more messages than cr50, so it can be difficult to debug cr50 console
issues if both consoles are set to debug.

BUG=none
BRANCH=none
TEST=verify setting cr50 loglevel does not affect the ec loglevel.

Change-Id: I529bf686443bf9df004e209e880f811608eef6a8
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1355967
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>