EFS: Jump to RW on AP shutdown
We need to re-run EFS when the system boots in recovery mode then
a power button is pressed.

This patch adds HOOK_CHIPSET_SHUTDOWN to trigger vboot_main.

NOTE: This may not be a good idea because in a hook context,
other tasks are running and may interfere with UART which is used
for Cr50 communication.

Fizz is fixed differently by resetting the system.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/119329144
BRANCH=none
TEST=Boot hatch

Change-Id: I110a499a158085b527abbe928d84b9f2ba00078f