Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> Hmm... I don't see a value in this patch. CONFIG_I2C_PASSTHRU_RESTRICTED is currently not defined and no plan to be defined.
> 
> If you want to prevent someone from defining it for Nami, you should have some sort of test, which runs ectool i2cread to get battery status. That'll be much more useful.

Hi Daisuke,

It is for sure that it is not currently defined, but don't know if others will in some future have plan to define so )

What I was hooping to address here is if someone accidentally defined the flag and we miss the chance to review it. At least this marco should cause a compile error so both side (in the transition to a more suitable solution) can be aware.

Sorry I didn't write that in the commit message. I have tried this in the ectool and confirmed that it gets the same value.

Patch-set: 3
Reviewer: Gerrit User 1001197 <1001197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>