chgstv2: fix LED glitch on reboot
this fixes a glitch where the LEDs are momentarily set to the battery
fail color when the EC initializes.  it takes a bit of time to confirm
battery presense - we were treating this window of uncertainty as a
battery failure.  there's another case where we try to wake up a
depleted battery (ST_PRECHARGE).  this should also not be treated as a
battery failure until the state machine moves to ST_IDLE after
PRECHARGE_TIMEOUT.

BUG=b:120200655
BRANCH=none
TEST=EC reboot no longer flashes the LEDs red

Change-Id: Ib85fc7282c3edb68d2e22b5d1f3071a89bbcb21d
Signed-off-by: Caveh Jalali <caveh@chromium.org>