PD 3.0: Update message type field size
For PD 3.0, bit 4 of the message header was added to the message type
field.  It should be safe to check that bit in all of our header type
checks regardless of version, since in PD 2.0 that field was
"Reserved/shall be set to 0".

BUG=None
BRANCH=None
TEST=builds, ampton PD negotiation works

Change-Id: I374c3d71210652a138e2ad53bd12e4b31eae9b0c
Signed-off-by: Diana Z <dzigterman@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/1363814
(cherry picked from commit fa953e4a856183600e3c32facbc33605fcc4836b)