Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If rowan (and kevin) are dead boards why do we keep the overlays in the repo and worry about building them?

kevin is definitely not dead. We still build that EC firmware on a branch, it's just that maintaining it on ToT has been particularly painful (frequently out of flash space), so we decided to remove it.

Patch-set: 1
Reviewer: Gerrit User 1129199 <1129199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>