Update patch set 1
Patch Set 1:

I am not sure, but I need a place to put the GPIO changes in order to flash the FP MCU for ekko.  I cannot change this in tot because this is a soft link to nocturne_fp, so it would break nocturne.  My solution is to pull the initial creation of nami_fp over to the nami fw branch and update flash_fp_mcu over there with the ekko gpio changes.  What do you think?

Patch-set: 1
Reviewer: Gerrit User 1117900 <1117900@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>