Update patch set 2
Patch Set 2: Code-Review+2

I am not going to argue, but let's keep this in mind for the future: it is much easier to keep track of what a change does when cosmetic and functional changes are separate.

It is very easy to prepare a patch which just changes names, comments, etc., and easy to review too.

And then it would be much easier to see what actually the fix is.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2