nami_fp: remove flash_fp_mcu file soft link
We need to change this soft link to a new file.  Somehow it's not
working when I try to do it in one go, so have a separate CL where we
remove the soft link with a follow up CL where we add the new file.
When I view the CL with gitk, I see the proper diffs, but seems like
there's a bug in gerrit that prevents the diffs from showing up.

BUG=b:119565385
BRANCH=None
TEST=None

Change-Id: I89a3978d1b081e7912d52c1688df5e18e45b1a85
Signed-off-by: Shelley Chen <shchen@google.com>