Update patch set 3
Patch Set 3:

OK to unblock for now but we need to productise this. While we're working on that, can you at least add an unmissable log message saying that this is updating a nami board? In case it ends up floating around and being use on another device

Patch-set: 3
Reviewer: Gerrit User 1158184 <1158184@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>