Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> I'm jumping in without all the context, so feel free to ignore, but since the entire script is shared except for the GPIO settings (GPIO_NRST, GPIO_BOOT0, GPIO_PWREN), could we just move that code into a board-specific config.sh (i.e., board/nami_fp/flash_fp_mcu_config.sh and board/nocturne_fp/flash_fp_mcu_config.sh) and then source that in a shared flash_fp_mcu script that is common to all? (Only makes sense if we don't expect any other divergence between the rest of the update script)

Yes, but I'm assuming that this is the follow up work that Nicolas (?) is going to do to product-ising this flash process?  This was a quick fix to get this script published for the factory build going on at the moment.

Patch-set: 4
Reviewer: Gerrit User 1117900 <1117900@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>