core/cortex-m: Support prevent chip memory region from GC.
We would like to keep a symbol in a chip memory region from GC in link
time.  However __attribute__((used)) cannot fulfill the requirement in such
situation.  This CL adds a "name.keep" section to prevent all the symbols
in this sction and in a chip memory region from GC.  Also, we would like to
support a non-NOLOAD section, which can load default value on runtime.

BUG=b:120825336
TEST=make buildall -j
BRANCH=none

Change-Id: I76cf445f6b4c0b61c20182a1aaf5a44f962049ae
Signed-off-by: Yilun Lin <yllin@google.com>