EFS: Send rollback version to Cr50
This patch makes EC send a rollback version to Cr50. Cr50 is supposed
to accept the rollback version if it's equal to or newwer than the
one saved in the TPM or to reject it otherwise.

A rollback version is specified by CONFIG_ROLLBACK_VERSION. It is
currently zero for all boards and can be independently incremented
by each board.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/119329144,chromium:998135
BRANCH=none
TEST=Boot Nami

Change-Id: I1377ca42feb62f85be84e0c71b8b248bd566d492