tcpc: wait for TCPC wake up upon first access
Previously we tried to perform i2c communication with the TCPC when it
was in low power mode and only if the i2c transaction failed did we wait
for the device to wake up. Now that the PS8751 will respond to i2c
transaction (ACK them) when it is in LPM, the initialization happens a
little later.

BRANCH=none
BUG=b:118063849
TEST=verify that plug and unplug for PS8751 does not get in a bad state
any more.

Verify that ANX3429, and ANX3447 LPM still works too.

Change-Id: I8c18195e55ee6d04af7d4ff24230a3bd2d147d53
Signed-off-by: Jett Rink <jettrink@chromium.org>