Update patch set 1
Patch Set 1:

What was the original intent of "Check registers to see if we can tell that the TCPC has reset."? If it was to handle unexpected cases of the TCPC resetting underneath us, then it seems like we'd want to keep this check even for the LPM case.

Patch-set: 1
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>