Update patch set 1
Patch Set 1: Code-Review+1

Two comments, of minor note:

> Original description:
> 
> According to the PD spec, in a source initiated Power Role swap, the
> initial source must change its termination from Rp to Rd prior to
> sending the PS_RDY control message.

The timing of applying Rd technically isn't specified in the spec. Merely that it happens after GoodCRC-to-Accept and before PS_RDY#1. I would instead describe this as [a workaround to address quirks with non-spec-compliant devices that "jump the gun" in applying their Rp before PS_RDY#2].

This patch isn't a spec requirement: it's a quirk + IOP fix, that happens to be spec-compliant to do.




> Patch Set 1:
> 
> I have tried this change previously, and it did not give the desired results. See the BUG= discussion for more info

I have reviewed the traces and behavior firsthand. I believe this may highlight exactly said "non-spec-compliant devices".

Please see [4.5.2.2.5.2 Exiting from Attached.SNK State] in the Type-C v1.3 spec, quoted below:

> "A port that has entered into USB PD communications with the Source and has seen the CC
> voltage exceed vRd-USB may monitor the CC pin to detect cable disconnect in addition to
> monitoring VBUS.

> A port that is monitoring the CC voltage for disconnect (but is not in the process of a USB PD
> PR_Swap or USB PD FR_Swap) shall transition to Unattached.SNK within tSinkDisconnect
> after the CC voltage remains below vRd-USB for tPDDebounce."

Please note the important text regarding "using CC 0v for disconnect detection": "(but is not in the process of a USB PD PR_Swap or USB PD FR_Swap)". 

As for the converse behavior we are fixing (seeing CC line spike to 3.3v vOpen briefly), please see [4.5.2.2.9.2 Exiting from Attached.SRC State] in the Type-C v1.3 spec:

> A Source that is not supplying VCONN shall transition to Unattached.SRC when the SRC.Open
> state is detected on the monitored CC pin. The Source shall detect the SRC.Open state within
> tSRCDisconnect, but should detect it as quickly as possible


Please not the important text about "how long" we need to see vOpen: "within tSRCDisconnect, but should detect it as quickly as possible".

This means CC touching vOpen even for a picosecond is enough to signal a disconnect -- and we are free to (and supposed to!!) act on that by kicking off the device. [Note: many high-performance FR_SWAP capable PD controllers I've tested *do exactly this*. A 5ps spike due to a non-compliant device inductively spiking the CC line was triggering a disconnect in its state machine.]


Based on what you verbally describe (but I need to scope and confirm), and the notes above -- the product you describe as "not working" with this patch is noncompliant. It is likely ignoring these hard requirements in the spec.

Patch-set: 1
Reviewer: Gerrit User 1269501 <1269501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1