Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Thank you for the detailed explanation Nitin. I guess we have to chose which devices to support because there are issues on both sides.
> 
> I do think the "bad stuff" for the monitor is that the display doesn't work (I would have to verify again though); where as the "bad stuff" for the qualcom stack is that the PD negotiation needs to happen again.
> 
> Even though we are being nicer to compliant devices, I still have reservation about making a change that we know breaks a monitor and doesn't improve another (customer perceived) broken scenario.
> 
> Let me verify what this change does again with the Dell monitor tomorrow and see if I am remembering the "bad stuff" correctly. Hopefully it will just work with this change though.

I apologize for bordering on TMI. It's more for whoever stumbles across this post in 10 years rather than saying "NVM, fixed it". :)

I would like to issue one strong philosophical caution, regarding your first bullet point: we should NOT behave noncompliantly, in order to work around problems in noncompliant products. I'm sure Benson would agree with me on that.

We should do what we can within the scope of still being compliant (like fudging timing) to try to work around these issues -- but if push comes to shove, I say we shove manufacturers to fix their products.

As the "stick", we should strictly enforce USB-IF compliance and Certification requirements. (And encourage Google not to buy products from noncompliant vendors.) As a "carrot", perhaps we can educate them about the spec and offer to guide them. I will see if it is possible to reach out to that particular product's manufacturer and point out this issue to them.

Sorry again for the tangeant, but thank you for bringing this point up. It is 100% valid and appropriate.

Patch-set: 2
Reviewer: Gerrit User 1269501 <1269501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>