g/i2cs: Sample SDA multiple times before considering the bus wedged
The tpm i2c protocol works as follows:
1. Master writes a 1 byte register address to the tpm
2. Master completes transaction then waits for GPIO_INT_AP_L to fire.
3. i2c write complete isr fires on the tpm
  1. register data is pushed into the tx fifo
  2. GPIO_INT_AP_L is pulsed to notify the master that the read is ready.
  3. tpm increments i2cs_read_irq_count
4. Master reads n bytes from the tpm.
5. Times passes
6a. Master decides to update a register on the TPM
  1. Master writes n bytes to the tpm.
  2. goto 2
6b. Master decides to read a register from the TPM
  1. goto 1

poll_read_state will currently poll SDA every 500 ms. If it sees that
there have been no write completions since the last polling period and
SDA is currently low it assumes the bus is wedged, and it resets the i2cs
controller. This logic has the potential to terminate in flight
transactions.

For example: the poller runs at step 4 and notices that a write
interrupt has occurred. It updates last_i2cs_read_irq_count. The master
waits 499.99 ms and then initiates a new transaction (1 or 6a.1).
The poller then runs again at 500ms and notices that the write complete
interrupt counter has not incremented and SDA is low because a 0 is
currently being transmitted on the bus. The poller then resets the i2cs
controller. This causes the master to receive a NAK.

This cl changes the logic so that the poller requires SDA to be low for
5 consecutive periods between write completes before it resets the
controller. This in theory could still cause a false positive. If there
was an i2c write followed by 5 i2c reads spaced 100ms apart. The tpm i2c
protocol doesn't function this way so don't have to worry about this.
The other false positive case is if there was a i2c write transaction
that took over 500ms, which is unlikely. This approach does suffer from
the fact that it can't detect if SDA is stuck high. If the master was
reading a 1 from the tpm or writing a 1 to the tpm and stopped clocking
the poller would never catch that and the i2cs controller would then be
wedged. We could partially fix this by looking at the read begin and
read complete interrupts and resetting if a transaction takes too long.
Unfortunately there is no write being interrupt so we can't tell how
long a write transaction takes to compute a timeout.

BUG=b:113880780
BRANCH=none
TEST=Ran a reboot stress test for over 16 hours and did not see any
transaction failures. Also manually grounded SDA to see if
read_recovery_count incremented:

[49.454601 Committing NVMEM changes.]
[49.478889 sda low count: 0]
[49.581901 sda low count: 0]
[49.682643 sda low count: 1]
[49.783386 sda low count: 2]
[49.884147 sda low count: 0]
[49.984863 sda low count: 0]
[50.085613 sda low count: 1]
[50.186359 sda low count: 2]
[50.287105 sda low count: 0]
[50.387847 sda low count: 0]
[50.488605 sda low count: 1]
[50.589355 sda low count: 2]
[50.690105 sda low count: 3]
[50.790859 sda low count: 4]
[50.891605 sda low count: 5]
i2c is hosed!

Change-Id: I7b6f446ee75b43e9d66a6a5e51dd077c60108f90
Signed-off-by: Raul E Rangel <rrangel@chromium.org>