Update patch set 1
Patch Set 1:

(1 comment)

I'm not crazy about how this is being done in this patch.  Really we want something that can be merged with ToT first, then pulled to the branch.  This patch obviously can't be merged to ToT, because it's in x86 generic code and will affect all platforms.

Even if we merged this to the firmware branch, it affects all grunt platforms, which doesn't seem to be needed.

Patch-set: 1
Reviewer: Gerrit User 1133034 <1133034@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>