ish: balance all __in_isr calls
If we encountered an unregistered IRQ vector we would not decrement
the __in_isr global properly which would cause the ISH to lock up. Also
we should be atomically incrementing and decrementing the __in_isr
counter otherwise we count end up in the same unbalanced scenario.

BRANCH=none
BUG=b:121343650,b:112750896
TEST=With patch stack on arcada (ISH5) 'taskinfo' command
works as expected.

Change-Id: I3975356226d92a81bfd207d77bba42f3f8b30bb8
Signed-off-by: Jett Rink <jettrink@chromium.org>