ish: fix 32-bit quotient overflow
Restructure the hw_clock_source_read to perform a divide by 12 operation
on a 64-bit number without causing a 32-bit overflow for the quotient by
performing the operation in stages. This is optimized specifically for
the value of 12, so we don't need to use the general purpose
uint64divmod method.

BRANCH=atlas
BUG=b:122314133
TEST=use fortime on atlas to put time next to rollover point and verify
that we no longer get a divide expection (#DE)

Change-Id: I0b5fe59bec9a703fddac657986106537d4e9e203
Signed-off-by: Jett Rink <jettrink@chromium.org>